% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{this function uses brute force as an algorithem}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame.}

\item{W}{A numeric. 

1}
}
\value{
the calculated value by this algorithem
}
\description{
this function uses brute force as an algorithem
}
