% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{This function uses greedy algorithem}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A data.frame.}

\item{W}{A numeric .}
}
\value{
it returns the calculated value and elements

  @export
}
\description{
This function uses greedy algorithem
}
