\name{knapsackalg-package}
\alias{knapsackalg-package}
\alias{knapsackalg}
\docType{package}
\title{
\packageTitle{knapsackalg}
}
\description{
\packageDescription{knapsackalg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsackalg}
\packageIndices{knapsackalg}
greedy_knapsack , knapsack_dynamic and 
}
\author{
\packageAuthor{knapsackalg}

Maintainer: \packageMaintainer{knapsackalg}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
