% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{"Lab 6 package"}
\alias{"Lab 6 package"}
\alias{knapsack_dynamic}
\title{Knapsack problem using dynamic algortihm}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A data.frame}

\item{W}{A number}
}
\value{
the maximum values and elements to be carried by knapsack
}
\description{
Knapsack problem using dynamic algortihm
}
