\name{knapsackalg-package}
\alias{knapsackalg-package}
\alias{knapsackalg}
\docType{package}
\title{
\packageTitle{knapsackalg}
}
\description{
\packageDescription{Algorithems for solving knapsack problem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsackalg}
\packageIndices{knapsackalg}
}
\author{
\packageAuthor{knapsackalg}

Maintainer: \packageMaintainer{knapsackalg}
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
