\name{knapsackalg-package}
\alias{knapsackalg-package}
\alias{knapsackalg}
\docType{package}
\title{
\packageTitle{knapsackalg}
}
\description{
\packageDescription{knapsackalg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{knapsackalg}
\packageIndices{knapsackalg}
greedy_knapsack , knapsack_dynamic and 
}
\author{
\packageAuthor{knapsackalg}

Maintainer: \packageMaintainer{knapsackalg}
}
\references{

}
\keyword{ package }
\seealso{

}
\examples{

}
